requirements.txt
amqp==5.0.1
asgiref==3.2.10
astroid==2.4.2
attrs==20.2.0
bcrypt==3.2.0
billiard==3.6.3.0
cached-property==1.5.2
celery==5.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.1.1
distro==1.5.0
Django==3.1.2
djangorestframework==3.12.1
docker==4.3.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
flake8==3.8.4
idna==2.10
isort==5.6.4
jsonschema==3.2.0
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
paramiko==2.7.2
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
wrapt==1.12.1
